Term 3, Week 3, August 10th
Forgot to upload changes for the past few days, what I have achieved is put the text into a spell check and then pasted it into the code on the Overfishing page. By using the single column layout that I have, the look is user-friendly and imitates an article of sorts. Advice given to me was that I should change up the page so it's not always the column otherwise it will be a bit repeatitive.

Thoughts about other pages have been flowing into my mind about either using similar layouts or entirely different (eg, Newspaper, multiple articles on one page, etc.)