Term 3, Week 6, Wednesday
Today I finished the oil spills page!!! essentailly what i have done is implemented the same Polaroid container where four polaroids sit in the solutions section of the page. Yes it is a copy and paste of the code, but the point of difference is the colour scheme of each image and background used. The effect this would have on the viewer is that it will look different showing a point of design difference.

I have also altered my media. There wasn't even a solutions text page on my media and now I have a full section of oil spill solutions! I have out it through no spell checks at all.

I have now started coral bleaching page where I will now copy and paste the code of the oil spills page and then alter that to make a completely different informational page.